+++ packages/adapter-chakra/src/DataTable.branches.test.tsx
@@ -14,8 +14,9 @@ import { ChakraProvider, defaultSystem } from "@chakra-ui/react";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 
-import { FilterDrawer, LoadingState } from "./components/chrome";
+import { FilterDrawer } from "./components/FilterDrawer";
 import { FilterPopover } from "./components/FilterPopover";
+import { LoadingState } from "./components/TableSkeleton";
 import { DataTable } from "./DataTable";
 import type { ColumnDef } from "./index";
